// src/rest-client-ts-node/generated/model/accuracyInput.ts
/**
 * Swagger Petstore
 * This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AccuracyInput {
    /**
    * number of True Positives
    */
    'tp'?: number;
    /**
    * number of False Positives
    */
    'fp'?: number;
    /**
    * number of False Negatives
    */
    'fn'?: number;
    /**
    * number of True Negatives
    */
    'tn'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "tp",
            "baseName": "tp",
            "type": "number"
        },
        {
            "name": "fp",
            "baseName": "fp",
            "type": "number"
        },
        {
            "name": "fn",
            "baseName": "fn",
            "type": "number"
        },
        {
            "name": "tn",
            "baseName": "tn",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return AccuracyInput.attributeTypeMap;
    }
}