// src/rest-client-ts-node/generated/model/correlationMatrixInput.ts
/**
 * Swagger Petstore
 * This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CorrelationMatrixInput {
    'x'?: Array<number>;
    'y'?: Array<number>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "x",
            "baseName": "x",
            "type": "Array<number>"
        },
        {
            "name": "y",
            "baseName": "y",
            "type": "Array<number>"
        }    ];

    static getAttributeTypeMap() {
        return CorrelationMatrixInput.attributeTypeMap;
    }
}